[or1k] Moves TLB Management to Cluster AL
In this commit, I move the TLB management to Cluster Abstraction
Layer but left elementary definitions and primitives for dealing
with the TLB entries.
